refactoring active orders
