+++ src/main/java/com/mediasoft/warehouse/configuration/RestConfigurationProperties.java
@@ -1,8 +1,8 @@
 package com.mediasoft.warehouse.configuration;
 
-import com.mediasoft.warehouse.configuration.service_properties.AccountServiceProperties;
-import com.mediasoft.warehouse.configuration.service_properties.CrmServiceProperties;
-import com.mediasoft.warehouse.configuration.service_properties.CurrencyServiceProperties;
+import com.mediasoft.warehouse.properties.AccountServiceProperties;
+import com.mediasoft.warehouse.properties.CrmServiceProperties;
+import com.mediasoft.warehouse.properties.CurrencyServiceProperties;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
